Server: Create and use shared types
