+++ server/src/controllers/availabilityControllers.ts
@@ -1,8 +1,6 @@
 import { type Request, type Response } from "express";
-import Availability, {
-  type IAvailability,
-  isAvailabilityList,
-} from "../models/Availability";
+import Availability, { isAvailabilityList } from "../models/Availability";
+import { type IAvailability } from "@shared/types/api/availability";
 import Reservation from "../models/Reservation";
 import Premise from "../models/Premise";
 import Space from "../models/Space";
